split unlock folders
