+++ shared/constants/reducer.js
@@ -18,7 +18,7 @@ import type {State as RouteTree} from '../constants/types/route-tree'
 import type {State as Settings} from '../constants/types/settings'
 import type {State as Signup} from '../constants/types/signup'
 import type {State as Tracker} from '../constants/types/tracker'
-import type {State as UnlockFolders} from '../constants/unlock-folders'
+import type {State as UnlockFolders} from '../constants/types/unlock-folders'
 import type {State as Waiting} from '../constants/waiting'
 
 export type TypedState = {
