+++ shared/constants/unlock-folders.js
@@ -1,21 +1,5 @@
 // @flow
-import type {DeviceID} from '../constants/types/flow-types'
-import type {DeviceType} from '../constants/types/devices'
-
-export type Device = {
-  type: DeviceType,
-  name: string,
-  deviceID: DeviceID,
-}
-
-export type State = {
-  closed: boolean,
-  devices: ?Array<Device>,
-  paperkeyError: ?string,
-  phase: 'dead' | 'promptOtherDevice' | 'paperKeyInput' | 'success',
-  sessionID: ?number,
-  waiting: boolean,
-}
+import type {State} from './types/unlock-folders'
 
 const initialState: State = {
   closed: true,
